+++ backend/app/core/config.py
@@ -1,7 +1,7 @@
 from pydantic_settings import BaseSettings
 from functools import lru_cache
 
-APP_VERSION = "0.0.7"
+APP_VERSION = "0.0.8"
 
 
 class Settings(BaseSettings):